refactor(routes): extract content and upload-waiver routes to content blueprint
Move six routes from _api.py into a new app/routes/content.py blueprint
mounted at /_api: server-time, stones, markdown/<slug>, render-markdown,
<tournament_url>/upload-waiver, and leagues/<league_url>/upload-waiver.
The URL surface is unchanged.

The upload-waiver routes were considered for waivers.py for topical
co-location, but that blueprint is currently unregistered and has no
/_api prefix, so moving the routes there would change their URLs.
Reactivating waivers.py is a separate concern.

_check_to and _require_league stay in _api.py for now and are imported
from there by content.py. This transitional import will dissolve when
later refactor stages move those helpers to permanent homes.

The render_markdown import is hoisted to the module top rather than
called inline, since app.filters does not depend on app.routes and the
inline form was unnecessary.
